Merge: 필드값 변경

See merge request s09-fintech-finance-sub2/S09P22B207!182

+++ Back-End/src/main/java/com/conseller/conseller/gifticon/dto/response/ExpiringGifticonResponse.java
@@ -10,5 +10,5 @@ public class ExpiringGifticonResponse {
     private Long userIdx;
     private Integer expiryDay;
     private String gifticonName;
-    private Integer gifticonCnt;
+    private Long gifticonCnt;
 }
